requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1,<3.0
reportlab>=4.0,<5.0
django-cors-headers>=3.14,<4.0